feat: fold AggregateMetrics into RunResult; delete aggregate.rs (#63)
`AggregateMetrics` was a thin wrapper over `&[MetricsSnapshot]` whose
methods were all "compute per-snapshot, then average across replicas."
Since `RunResult<C>` already owns `metrics: Vec<MetricsSnapshot>` and
`duration: Duration`, the wrapper was dead weight. Methods move to
`RunResult<C>`:

- `p50_latency_ms`, `p90_latency_ms`
- `leaders_committed_per_second`     (was `leader_committed_per_second`)
- `transactions_committed_per_second`

Bonus simplification: the rate methods drop their `Duration` param,
reading `self.duration` directly. Caller side is one fewer `let
duration = …` plus one less indirection.

Naming: pluralised `leader_committed_per_second` →
`leaders_committed_per_second` and named the new per-replica view
`leaders_committed_per_replica` (was `commit_count_per_replica`).
"Leaders" reads as the things being counted, "per_X" as the unit each
entry is keyed by — consistent across all three accessors and
grammatically natural.

`crates/dag/src/metrics/aggregate.rs` deleted; `mod aggregate` and
`pub use AggregateMetrics` removed from metrics.rs. Stale comment in
names.rs updated.

cli's terminal renderer (`cli/src/terminal/result.rs`) drops the
`AggregateMetrics::new(...)` ceremony and calls methods on `&self`
directly. Numerical output is bit-for-bit identical (averages don't
change shape; `* 1000` already moved into the per-snapshot helper in
an earlier commit).

Refs #63, tracked by #66.
